Create README.md
Gender, like any artistic expressions, is a performance. There lies the 
long-existing distinction between sex and gender that sex is biological whereas gender 
is culturally constructed. Butler argues that if the so-called “sex” is as culturally 
constructed as gender, perhaps “sex” has already been gendered, leading to the 
collapse of demarcation between sex and gender (Butler, “Critically Queer”, p. 7). 
With its substantive metaphysical discourse, gender has turned out to be a 
performance, constituting its gender identity (Butler, “Critically Queer”, p. 25). In 
Confession of a Mask, the male protagonist Mishima acts as a masculine man who is 
interested in women to conceal his homosexuality. Throughout different phases of his 
life, he attempts to conform to social norms by putting his “mask” on. The “mask” 
symbolizes his fake/counterfeit gender identity, the pretense of being a heterosexual. 
By confessing of the mask as the title suggests, he unveils/uncovers his homosexual 
identity that he does not dare to confront, resulting in melancholia. He feels guilty, 
painful, shameful, a heightened sense of loss. Torn between the boundaries between 
heterosexuality and homosexuality, he cannot stand the culmination of melancholia in 
which his “self” or subjectivity is pathologically “abjected”. The existence of “self” 
has become an enigma in the end.